Add EMG mainloop
The main loop serves as the guiding game logic for Simon Says.

The current file waits for muscle inputs and then outputs what muscles were activated to the LED display.

In future iterations, only the randomized sequence will be output to the LEDS while the muscle impulses will be compared on the computer to see if they matched the outputted sequence.